fix: clean up new prefs and fix npe

+++ Explorer/Assets/DCL/Prefs/FileDCLPlayerPrefs.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using System.Threading.Tasks;
 using UnityEngine;
 
 namespace DCL.Prefs
@@ -16,6 +17,8 @@ namespace DCL.Prefs
         private readonly UserData userData;
 
         private readonly UnityDCLPlayerPrefs unityPrefs;
+
+        private bool dataChanged;
 
         public FileDCLPlayerPrefs()
         {
@@ -45,20 +48,26 @@ namespace DCL.Prefs
 
         public void SetString(string key, string value)
         {
-            DeleteKey(key);
+            if (userData.Strings.TryGetValue(key, out string existing) && existing == value) return;
+
             userData.Strings[key] = value;
+            dataChanged = true;
         }
 
         public void SetInt(string key, int value)
         {
-            DeleteKey(key);
+            if (userData.Ints.TryGetValue(key, out int existing) && existing == value) return;
+
             userData.Ints[key] = value;
+            dataChanged = true;
         }
 
         public void SetFloat(string key, float value)
         {
-            DeleteKey(key);
+            if (userData.Floats.TryGetValue(key, out float existing) && Mathf.Approximately(existing, value)) return;
+
             userData.Floats[key] = value;
+            dataChanged = true;
         }
 
         public string GetString(string key, string defaultValue)
@@ -80,14 +89,17 @@ namespace DCL.Prefs
         }
 
         public bool HasKey(string key) =>
-            unityPrefs.HasKey(key) || userData.Strings.ContainsKey(key) || userData.Ints.ContainsKey(key) || userData.Floats.ContainsKey(key);
+            (unityPrefs != null && unityPrefs.HasKey(key)) || userData.Strings.ContainsKey(key) || userData.Ints.ContainsKey(key) || userData.Floats.ContainsKey(key);
 
         public void DeleteKey(string key)
         {
+            if (!HasKey(key)) return;
+
             userData.Strings.Remove(key);
             userData.Ints.Remove(key);
             userData.Floats.Remove(key);
             PlayerPrefs.DeleteKey(key);
+            dataChanged = true;
         }
 
         public void DeleteAll()
@@ -99,18 +111,29 @@ namespace DCL.Prefs
 
         public void Save()
         {
-            fileStream.Seek(0, SeekOrigin.Begin);
-            fileStream.SetLength(0);
+            if (!dataChanged) return;
 
-            using var writer = new StreamWriter(fileStream, Encoding.UTF8, 1024, true);
-            string json = JsonConvert.SerializeObject(userData);
-            writer.Write(json);
-            writer.Flush();
+            dataChanged = false;
+
+            // Run save on a background thread
+            // Task.Run(() =>
+            // {
+            //     lock (fileStream)
+            //     {
+                    fileStream.Seek(0, SeekOrigin.Begin);
+                    fileStream.SetLength(0);
+
+                    using var writer = new StreamWriter(fileStream, Encoding.UTF8, 1024, true);
+                    string json = JsonConvert.SerializeObject(userData);
+                    writer.Write(json);
+                    writer.Flush();
+            //     }
+            // });
         }
 
         private void MigrateString(string key)
         {
-            if (unityPrefs != null && !unityPrefs.HasKey(key)) return;
+            if (unityPrefs == null || !unityPrefs.HasKey(key)) return;
 
             userData.Strings.Add(key, unityPrefs!.GetString(key, null));
             unityPrefs.DeleteKey(key);
@@ -119,7 +142,7 @@ namespace DCL.Prefs
 
         private void MigrateInt(string key)
         {
-            if (unityPrefs != null && !unityPrefs.HasKey(key)) return;
+            if (unityPrefs == null || !unityPrefs.HasKey(key)) return;
 
             userData.Ints.Add(key, unityPrefs!.GetInt(key, 0));
             unityPrefs.DeleteKey(key);
@@ -128,7 +151,7 @@ namespace DCL.Prefs
 
         private void MigrateFloat(string key)
         {
-            if (unityPrefs != null && !unityPrefs.HasKey(key)) return;
+            if (unityPrefs == null || !unityPrefs.HasKey(key)) return;
 
             userData.Floats.Add(key, unityPrefs!.GetFloat(key, 0f));
             unityPrefs.DeleteKey(key);
